requirements.txt
paho-mqtt==1.5.0
numpy==1.16.1
scikit-learn==0.21.3
opencv-python==4.1.2.30
tensorflow==1.15.0
scipy==1.1.0
Pillow==7.2.0
pymongo==3.11.0
python-dotenv==0.15.0
imutils==0.5.3
requests
